+++ cyclos_configuration/scripts/emandates/Operation_InternalToggleBlock.groovy
@@ -1,6 +1,6 @@
 import org.cyclos.entities.users.User
 
-Map<String, String> scriptParameters = binding.scriptParameters
+Utils utils = new Utils(binding)
 Map<String, Object> formParameters = binding.formParameters
 User user = formParameters.user
 
@@ -12,15 +12,15 @@ def result = ''
 switch(lock) {
     case 'blocked':
         usr.emandates_lock = ''
-        result = scriptParameters["locking.unblocked"]
+        result = utils.dynamicMessage("emResultUnblocked")
         break
     case null:
         usr.emandates_lock = 'blocked'
-        result = scriptParameters["locking.blocked"]
+        result = utils.dynamicMessage("emResultBlocked")
         break
     default:
         // Do nothing and bail out. We should never get here, the user may only toggle between 'blocked' and empty.
-        return scriptParameters["locking.error"] + usr.emandates_lock
+        return "${utils.dynamicMessage("emResultError")} ${usr.emandates_lock}"
 }
 userService.save(usrDTO)